Relicensing Rhino under MPL, see:http://groups.google.com/group/mozilla.dev.tech.js-engine/browse_thread/thread/12b1279e97d1f8a

+++ src/org/mozilla/javascript/NativeJavaObject.java
@@ -1,41 +1,46 @@
 /* -*- Mode: java; tab-width: 8; indent-tabs-mode: nil; c-basic-offset: 4 -*-
  *
- * The contents of this file are subject to the Netscape Public
- * License Version 1.1 (the "License"); you may not use this file
- * except in compliance with the License. You may obtain a copy of
- * the License at http://www.mozilla.org/NPL/
+ * ***** BEGIN LICENSE BLOCK *****
+ * Version: MPL 1.1/GPL 2.0/LGPL 2.1
  *
- * Software distributed under the License is distributed on an "AS
- * IS" basis, WITHOUT WARRANTY OF ANY KIND, either express or
- * implied. See the License for the specific language governing
- * rights and limitations under the License.
+ * The contents of this file are subject to the Mozilla Public License Version
+ * 1.1 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ * http://www.mozilla.org/MPL/
+ *
+ * Software distributed under the License is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
+ * for the specific language governing rights and limitations under the
+ * License.
  *
  * The Original Code is Rhino code, released
  * May 6, 1999.
  *
- * The Initial Developer of the Original Code is Netscape
- * Communications Corporation.  Portions created by Netscape are
- * Copyright (C) 1997-2000 Netscape Communications Corporation. All
- * Rights Reserved.
+ * The Initial Developer of the Original Code is
+ * Netscape Communications Corporation.
+ * Portions created by the Initial Developer are Copyright (C) 1997-2000
+ * the Initial Developer. All Rights Reserved.
  *
  * Contributor(s):
- * Norris Boyd
- * Igor Bukanov
- * Frank Mitchell
- * Mike Shaver
- * Kemal Bayram
+ *   Norris Boyd
+ *   Igor Bukanov
+ *   Frank Mitchell
+ *   Mike Shaver
+ *   Kemal Bayram
  *
- * Alternatively, the contents of this file may be used under the
- * terms of the GNU Public License (the "GPL"), in which case the
- * provisions of the GPL are applicable instead of those above.
- * If you wish to allow use of your version of this file only
- * under the terms of the GPL and not to allow others to use your
- * version of this file under the NPL, indicate your decision by
- * deleting the provisions above and replace them with the notice
- * and other provisions required by the GPL.  If you do not delete
- * the provisions above, a recipient may use your version of this
- * file under either the NPL or the GPL.
- */
+ * Alternatively, the contents of this file may be used under the terms of
+ * either the GNU General Public License Version 2 or later (the "GPL"), or
+ * the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
+ * in which case the provisions of the GPL or the LGPL are applicable instead
+ * of those above. If you wish to allow use of your version of this file only
+ * under the terms of either the GPL or the LGPL, and not to allow others to
+ * use your version of this file under the terms of the MPL, indicate your
+ * decision by deleting the provisions above and replace them with the notice
+ * and other provisions required by the GPL or the LGPL. If you do not delete
+ * the provisions above, a recipient may use your version of this file under
+ * the terms of any one of the MPL, the GPL or the LGPL.
+ *
+ * ***** END LICENSE BLOCK ***** */
 
 package org.mozilla.javascript;
 
@@ -707,8 +712,8 @@ WrapFactory#wrap(Context cx, Scriptable scope, Object obj, Class)}
             }
             return new Character((char)toInteger(value,
                                                  ScriptRuntime.CharacterClass,
-                                                 Character.MIN_VALUE,
-                                                 Character.MAX_VALUE));
+                                                 (double)Character.MIN_VALUE,
+                                                 (double)Character.MAX_VALUE));
         }
 
         // Double, Float
@@ -731,10 +736,10 @@ WrapFactory#wrap(Context cx, Scriptable scope, Object obj, Class)}
                 }
                 else {
                     double absNumber = Math.abs(number);
-                    if (absNumber < Float.MIN_VALUE) {
+                    if (absNumber < (double)Float.MIN_VALUE) {
                         return new Float((number > 0.0) ? +0.0 : -0.0);
                     }
-                    else if (absNumber > Float.MAX_VALUE) {
+                    else if (absNumber > (double)Float.MAX_VALUE) {
                         return new Float((number > 0.0) ?
                                          Float.POSITIVE_INFINITY :
                                          Float.NEGATIVE_INFINITY);
@@ -754,8 +759,8 @@ WrapFactory#wrap(Context cx, Scriptable scope, Object obj, Class)}
             else {
                 return new Integer((int)toInteger(value,
                                                   ScriptRuntime.IntegerClass,
-                                                  Integer.MIN_VALUE,
-                                                  Integer.MAX_VALUE));
+                                                  (double)Integer.MIN_VALUE,
+                                                  (double)Integer.MAX_VALUE));
             }
         }
 
@@ -787,8 +792,8 @@ WrapFactory#wrap(Context cx, Scriptable scope, Object obj, Class)}
             else {
                 return new Short((short)toInteger(value,
                                                   ScriptRuntime.ShortClass,
-                                                  Short.MIN_VALUE,
-                                                  Short.MAX_VALUE));
+                                                  (double)Short.MIN_VALUE,
+                                                  (double)Short.MAX_VALUE));
             }
         }
 
@@ -799,8 +804,8 @@ WrapFactory#wrap(Context cx, Scriptable scope, Object obj, Class)}
             else {
                 return new Byte((byte)toInteger(value,
                                                 ScriptRuntime.ByteClass,
-                                                Byte.MIN_VALUE,
-                                                Byte.MAX_VALUE));
+                                                (double)Byte.MIN_VALUE,
+                                                (double)Byte.MAX_VALUE));
             }
         }
